services/auth_service.py
"""Auth service for Google OAuth (optional)."""

import logging
from google.oauth2 import service_account

logger = logging.getLogger(__name__)


class AuthService:
    """Service for handling authentication and token refresh."""
    
    def __init__(self):
        """Initialize auth service."""
        pass
    
    def refresh_token(self, credentials):
        """Refresh expired credentials."""
        try:
            from google.auth.transport.requests import Request
            credentials.refresh(Request())
            return credentials
        except Exception as e:
            logger.error(f"Error refreshing token: {e}")
            raise
