requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
google-auth==2.25.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
